request-benchmark: split main.go into per-mode and helper files
